requirements.txt
pandas>=2.0,<3.0
numpy>=1.24,<3.0
scipy>=1.10,<2.0
pyserial>=3.5,<4.0
matplotlib>=3.7,<4.0
